Finalize extension split.
