+++ docs/src/literate/distributed_assembly_hypre.jl
@@ -14,64 +14,16 @@
 #md # The full program, without comments, can be found in the next [section](@ref heat_equation-plain-program).
 #
 # First we load Ferrite, and some other packages we need
-using Ferrite, MPI
-using HYPRE
-using PartitionedArrays, Metis
+using Ferrite
+using MPI, Metis # Loads FerriteMPI 
+using HYPRE # Loads FerriteHYPRE
 
 # Launch MPI
 MPI.Init()
 HYPRE.Init()
 
-function Ferrite.create_sparsity_pattern(::Type{<:HYPREMatrix}, dh::DofHandler, ch::Union{ConstraintHandler,Nothing}=nothing; kwargs...)
-    K = create_sparsity_pattern(dh, ch; kwargs...)
-    fill!(K.nzval, 1)
-    return HYPREMatrix(K)
-end
-
-###########################################
-## HYPREAssembler and associated methods ##
-###########################################
-
-struct HYPREAssembler <: Ferrite.AbstractSparseAssembler
-    A::HYPRE.HYPREAssembler
-end
-
-Ferrite.matrix_handle(a::HYPREAssembler) = a.A.A.A # :)
-Ferrite.vector_handle(a::HYPREAssembler) = a.A.b.b # :)
-
-function Ferrite.start_assemble(K::HYPREMatrix, f::HYPREVector)
-    return HYPREAssembler(HYPRE.start_assemble!(K, f))
-end
-
-function Ferrite.assemble!(a::HYPREAssembler, dofs::AbstractVector{<:Integer}, ke::AbstractMatrix, fe::AbstractVector)
-    HYPRE.assemble!(a.A, dofs, ke, fe)
-end
-
-function Ferrite.end_assemble(a::HYPREAssembler)
-    HYPRE.finish_assemble!(a.A)
-end
-
-## Methods for arrayutils.jl ##
-
-function Ferrite.addindex!(A::HYPREMatrix, v, i::Int, j::Int)
-    nrows = HYPRE_Int(1)
-    ncols = Ref{HYPRE_Int}(1)
-    rows = Ref{HYPRE_BigInt}(i)
-    cols = Ref{HYPRE_BigInt}(j)
-    values = Ref{HYPRE_Complex}(v)
-    HYPRE.@check HYPRE_IJMatrixAddToValues(A.ijmatrix, nrows, ncols, rows, cols, values)
-    return A
-end
-
-function Ferrite.addindex!(b::HYPREVector, v, i::Int)
-    nvalues = HYPRE_Int(1)
-    indices = Ref{HYPRE_BigInt}(i)
-    values = Ref{HYPRE_Complex}(v)
-    HYPRE.@check HYPRE_IJVectorAddToValues(b.ijvector, nvalues, indices, values)
-    return b
-end
-
 FerriteMPI = Base.get_extension(Ferrite, :FerriteMPI)
+FerriteHYPRE = Base.get_extension(Ferrite, :FerriteHYPRE)
 
 # We start generating a simple grid with 20x20 quadrilateral elements
 # and distribute it across our processors using `generate_distributed_grid`. 
@@ -181,14 +133,9 @@ precond = HYPRE.BoomerAMG()
 solver = HYPRE.PCG(; Precond = precond)
 uh = HYPRE.solve(solver, K, f)
 
-# TODO how to put this into an interface.
-# Copy solution from HYPRE to Julia
-uj = Vector{Float64}(undef, FerriteMPI.num_local_true_dofs(dh))
-copy!(uj, uh)
-
 # And convert from HYPRE to Ferrite
 u_local = Vector{Float64}(undef, FerriteMPI.num_local_dofs(dh))
-FerriteMPI.hypre_to_ferrite!(u_local, uj, dh)
+FerriteHYPRE.hypre_to_ferrite!(u_local, uh, dh)
 
 # # ### Exporting via PVTK
 # # To visualize the result we export the grid and our field `u`
