+++ ext/FerriteMPI.jl
@@ -5,7 +5,7 @@ module FerriteMPI
 
 using Ferrite
 # TODO remove me. These are merely hotfixes to split the extensions trasiently via an internal API.
-import Ferrite: getglobalgrid, num_global_dofs, num_global_dofs, num_local_dofs, num_local_true_dofs, global_comm, interface_comm, global_rank, compute_owner, local_dof_range
+import Ferrite: getglobalgrid, num_global_dofs, num_global_dofs, num_local_dofs, num_local_true_dofs, global_comm, interface_comm, global_rank, compute_owner, local_dof_range, remote_entities
 using Metis
 using MPI
 using Base: @propagate_inbounds
@@ -14,5 +14,9 @@ include("FerriteMPI/grid.jl")
 include("FerriteMPI/DistributedDofHandler.jl")
 include("FerriteMPI/iterators.jl")
 include("FerriteMPI/vtk-export.jl")
-    
+
+function __init__()
+    @info "FerriteMPI extension loaded."
+end
+
 end # module FerriteMPI
